Make QT debugger file names match class names without redundant prefix
