+++ src/osd/modules/debugger/qt/deviceswindow.h
@@ -3,7 +3,7 @@
 
 #include <QtGui/QtGui>
 
-#include "debugqtwindow.h"
+#include "windowqt.h"
 
 
 //============================================================
