+++ src/osd/modules/debugger/qt/windowqt.h
@@ -1,5 +1,5 @@
-#ifndef __DEBUG_QT_WINDOW_H__
-#define __DEBUG_QT_WINDOW_H__
+#ifndef __DEBUG_QT_WINDOW_QT_H__
+#define __DEBUG_QT_WINDOW_QT_H__
 
 #include <QtGui/QtGui>
